+++ pyproject.toml
@@ -38,7 +38,7 @@ ignore_errors = true
 
 [tool.semantic_release]
 branch = "main"
-version_variable = "gitlab/__version__.py:__version__"
+version_variable = "gitlab/_version.py:__version__"
 commit_subject = "chore: release v{version}"
 commit_message = ""
